Completed test outine for weekly price quotes.
Modified daily/weekly price quotes to exclude the retrieval of
today's date if today is a weekday before 9 PM EST. This is to ensure that we don't
write partial data for today since trading is ongoing. 9 PM repesents one hour past
the close of extended trading. The one hour delay is to ensure that accurate
daily data has been written to the data source.